teach authentication and authorization

+++ note_taker/notes/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 
 from .models import Note
 from .forms import NoteForm, SearchForm, FilterForm
@@ -32,19 +34,27 @@ def note_detail(request, note_pk):
     note = get_object_or_404(Note, pk=note_pk)
     return render(request, "notes/note.html", {'note': note})
 
+
+@login_required
 def add_note(request):
     form = NoteForm()
     if request.method == "POST":
         form = NoteForm(request.POST)
         if form.is_valid():
-            form.save()
+            note = form.save(commit=False)
+            note.added_by = request.user
+            note.save()
             return redirect("notes:note_list")
     context = {'form': form}
     return render(request, "notes/add-note.html", context)
 
 
+@login_required
 def update_note(request, note_id):
     note = get_object_or_404(Note, pk=note_id)
+    if request.user != note.added_by:
+        messages.error(request, "You do not have permission to perform that action")
+        return redirect("notes:note_list")
     form = NoteForm(instance=note)
     if request.method == "POST":
         form = NoteForm(request.POST, instance=note)
@@ -54,13 +64,20 @@ def update_note(request, note_id):
     context = {'note': note, 'form': form}
     return render(request, "notes/edit-note.html", context)
 
-
+@login_required
 def confirm_delete(request, note_pk):
     note = get_object_or_404(Note, pk=note_pk)
+    if request.user != note.added_by:
+        messages.error(request, "You do not have permission to perform that action")
+        return redirect("notes:note_list")
     return render(request, "notes/confirm-delete.html", context={"note": note})
 
+@login_required
 def delete_note(request, note_id):
     note = get_object_or_404(Note, pk=note_id)
+    if request.user != note.added_by:
+        messages.error(request, "You do not have permission to perform that action")
+        return redirect("notes:note_list")
     if request.method == "POST":
         note.delete()
     return redirect("notes:note_list")
